python/requirements.txt
sqlalchemy>=2.0
pymysql>=1.1
scikit-learn>=1.4
numpy>=1.26
ultralytics==8.3.85
pillow
fastapi==0.115.8
uvicorn[standard]==0.27.1
python-multipart>=0.0.9
